requirements.txt
# ══════════════════════════════════════════════════════════
# Urban Safety AI Desktop — requirements.txt
# Python 3.9+  |  Windows / macOS / Linux
# ══════════════════════════════════════════════════════════

# ── GUI framework ─────────────────────────────────────────
PyQt6>=6.5.0

# ── Deep learning ────────────────────────────────────────
# For GPU (CUDA 11.8):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For GPU (CUDA 12.1):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# For CPU only (universal):
torch>=2.0.0
torchvision>=0.15.0

# ── Computer vision ──────────────────────────────────────
opencv-python>=4.8.0
Pillow>=10.0.0

# ── Audio alerts (optional — Windows uses winsound built-in)
playsound>=1.3.0      ; sys_platform != "win32"

# ── Utilities ────────────────────────────────────────────
numpy>=1.24.0

# ── Build executable (optional) ─────────────────────────
pyinstaller>=6.0.0
